streamlit_app/requirements.txt
# Streamlit
streamlit==1.31.0

# Reddit Scraping
praw==7.7.1

# YouTube Scraping
google-api-python-client==2.116.0
youtube-transcript-api==0.6.2

# Blog Scraping
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.31.0

# OpenAI / Brand Voice Analysis
openai>=1.54.0

# Utilities
python-dotenv==1.0.1

# Social Media Posting
tweepy==4.14.0
Mastodon.py==1.8.1
